+++ server/photomory/src/main/java/com/example/photomory/entity/Comment.java
@@ -11,7 +11,7 @@ public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "comment_id")
-    private Long commentId;  // ✅ 고유 ID 필드
+    private Integer commentId;  // ✅ 고유 ID 필드
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "album_id", nullable = false)
@@ -40,11 +40,11 @@ public class Comment {
     }
 
     // ✅ Getters & Setters
-    public Long getCommentId() {
+    public Integer getCommentId() {
         return commentId;
     }
 
-    public void setCommentId(Long commentId) {
+    public void setCommentId(Integer commentId) {
         this.commentId = commentId;
     }
